# services/api-gateway/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "api-gateway",
    "version": "1.0.0",
    "description": "Puerta pública del sistema. Reenvía el path, el query, el cuerpo y el código de estado. GET /health no consulta backends. Si un destino no responde, el borde responde 503 JSON. El Swagger de cada servicio vive en su red interna; este documento es el contrato que se llama desde el host."
  },
  "security": [{ "BearerAuth": [] }],
  "paths": {
...
    "/auth/login": {
      "post": {
        "security": [],
        "summary": "Autenticar y obtener un access JWT",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["usuario", "contrasena"],
                "properties": {
                  "usuario": { "type": "string", "example": "admin" },
                  "contrasena": { "type": "string", "format": "password", "example": "Admin1234!" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "JWT de acceso" },
          "401": { "description": "Credenciales inválidas o cuenta inactiva" },
          "503": { "$ref": "#/components/responses/NoDisponible" }
        }
      }
    },
    "/auth/recover-password": {
      "post": {
        "security": [],
        "summary": "Solicitar recuperación de contraseña",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email"],
                "properties": { "email": { "type": "string", "format": "email" } }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Respuesta neutra exista o no el correo" },
          "503": { "$ref": "#/components/responses/NoDisponible" }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "security": [],
        "summary": "Establecer contraseña con un reset token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["token", "contrasena"],
                "properties": {
                  "token": { "type": "string" },
                  "contrasena": { "type": "string", "format": "password" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Contraseña actualizada" },
          "400": { "description": "Token, política o estado inválido" },
          "503": { "$ref": "#/components/responses/NoDisponible" }
        }
      }
    },
    "/auth/change-password": {
      "post": {
        "summary": "Cambiar la contraseña del usuario autenticado",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["contrasenaActual", "contrasenaNueva"],
                "properties": {
                  "contrasenaActual": { "type": "string", "format": "password" },
                  "contrasenaNueva": { "type": "string", "format": "password" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Contraseña actualizada" },
          "400": { "description": "Contraseña actual o política inválida" },
          "401": { "description": "JWT ausente o inválido" },
          "503": { "$ref": "#/components/responses/NoDisponible" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access JWT retornado por POST /auth/login"
      }
    },
    "responses": {
      "NoDisponible": {
        "description": "El backend no respondió",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/NoDisponible" } } }
      }
    },
    "schemas": {
...
      "NoDisponible": {
        "type": "object",
        "properties": {
          "status": { "type": "integer", "example": 503 },
          "mensaje": { "type": "string" },
          "servicio": { "type": "string" }
        }
      },
...
    }
  }
}